kafka_clickhouse_etl/requirements.txt
kafka-python==2.0.2
clickhouse_driver==0.2.9